fix warning

+++ examples/c_render/main.c
@@ -47,7 +47,7 @@ int main() {
 	// to view the changes live open render.ppm in an image viewer
 	for (size_t i = 0; i < 10; i++) {
 		unsigned char* image_buffer = render_refine(5);
-		printf("Step %d/10: Saving to 'render_c.ppm'...\n", i + 1);
+		printf("Step %d/10: Saving to 'render_c.ppm'...\n", ((int)i + 1));
 		save_image_as_ppm("render_c.ppm", image_buffer, width, height);
 	}
